third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-scrollbar-mainframe-expected.txt

This tests scroll gesture events on main frame scroll bars. The document should be slightly scrolled down if successful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.innerWidth - document.body.clientWidth is > 5
PASS window.scrollY is 800
PASS window.scrollY is 800
PASS window.scrollY is > 820
PASS window.scrollY is > 885
PASS window.scrollY is > 885
PASS successfullyParsed is true

TEST COMPLETE
